Factor out LinearGradient and RadialGradient
