+++ packages/flutter/lib/painting.dart
@@ -27,6 +27,7 @@ export 'src/painting/decoration.dart';
 export 'src/painting/edge_insets.dart';
 export 'src/painting/flutter_logo.dart';
 export 'src/painting/fractional_offset.dart';
+export 'src/painting/gradient.dart';
 export 'src/painting/text_painter.dart';
 export 'src/painting/text_span.dart';
 export 'src/painting/text_style.dart';
